+++ evm-block-execution/src/transaction/types/eip1559.rs
@@ -2,7 +2,7 @@
 //!
 //! [EIP-1559]: https://eips.ethereum.org/EIPS/eip-1559
 
-use super::{
+use super::codec::{
     TxDecodeError, append_access_list, append_destination, append_signature, decode_access_list,
     decode_destination, decode_signature, expect_items,
 };
